+++ eyes.images.java/src/test/java/com/applitools/eyes/images/TestConfiguration.java
@@ -1,5 +1,6 @@
 package com.applitools.eyes.images;
 
+import com.applitools.eyes.config.Configuration;
 import com.applitools.eyes.utils.ReportingTestSuite;
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -11,7 +12,7 @@ public class TestConfiguration extends ReportingTestSuite {
     }
 
     @Test
-    public void TestSetEnablePatterns() {
+    public void testSetEnablePatterns() {
         Eyes e = new Eyes();
 
         e.setEnablePatterns(true);
@@ -30,5 +31,24 @@ public class TestConfiguration extends ReportingTestSuite {
         Assert.assertEquals(eyes.getConfiguration().getMatchTimeout(), originalMatchTimeout);
         eyes.getConfigurationInstance().setMatchTimeout(newMatchTimeout);
         Assert.assertEquals(eyes.getConfiguration().getMatchTimeout(), newMatchTimeout);
+    }
+
+    @Test
+    public void testConfiguration() {
+        Configuration config = new Configuration();
+
+        Eyes e1 = new Eyes();
+        e1.setConfiguration(config.setAppName("app1"));
+
+        Eyes e2 = new Eyes();
+        e2.setConfiguration(config.setAppName("app2"));
+
+        Eyes e3 = new Eyes();
+        config.setAppName("DefaultAppName");
+        e3.setConfiguration(config);
+
+        Assert.assertEquals("app1", e1.getConfiguration().getAppName(), "e1 app name");
+        Assert.assertEquals("app2", e2.getConfiguration().getAppName(), "e2 app name");
+        Assert.assertEquals("DefaultAppName", e3.getConfiguration().getAppName(), "e3 app name");
     }
 }